+++ test-suite-azure-function-http-context-path/src/test/java/io/micronaut/azure/DemoController.java
@@ -1,0 +1,16 @@
+package io.micronaut.azure;
+
+import io.micronaut.http.annotation.*;
+import io.micronaut.http.annotation.Produces;
+import io.micronaut.http.MediaType;
+
+@Controller("/demo")
+public class DemoController {
+
+    @Produces(MediaType.TEXT_PLAIN)
+    @Get
+    public String index() {
+        return "Example Response";
+    }
+}
+
